Time: 2 ms (59.08%) | Memory: 7.3 MB (69.47%) - LeetSync

+++ 367-valid-perfect-square/valid-perfect-square.cpp
@@ -4,7 +4,7 @@ public:
         if( num <2){
             return true;
         }
-        long left = 2 , right = num/2, guess_squared, mid;
+        long long left = 2 , right = num/2, guess_squared, mid;
         while(left <= right){
             mid = (left + right)/2;
             guess_squared = mid*mid;
